+++ CHANGELOG.md
@@ -21,6 +21,8 @@
   - _selected menu-item isn't into view sometimes with 1st opening popup_
 - **Combobox controls (Select, Date, Time)**. _Popup shows/hides when user select text with mouseUp on label (outside input)_
 - [CalendarControl](src/controls/calendar.ts). _`$value=undefined` doesn't reset selected_
+- [NumberControl](src/controls/number.ts).
+  - _Alt+MouseWheel changes +-0.1 when decimal isn't allowed_
 
 **New/Features**: